Moved Core Commands to designated folder and added timer(task) checker
