+++ src/main/java/com/rosekingdom/rosekingdom/Core/CommandManager/CommandManager.java
@@ -1,7 +1,12 @@
 package com.rosekingdom.rosekingdom.Core.CommandManager;
 
 import com.rosekingdom.rosekingdom.Core.CommandManager.ExampleCommands.test;
-import com.rosekingdom.rosekingdom.Core.CommandManager.FeatureTesting.*;
+import com.rosekingdom.rosekingdom.Core.CommandManager.FeatureTesting.Demoscreen;
+import com.rosekingdom.rosekingdom.Core.CommandManager.FeatureTesting.ProtocolTest;
+import com.rosekingdom.rosekingdom.Core.Commands.CheckPermissions;
+import com.rosekingdom.rosekingdom.Core.Commands.PlayTime;
+import com.rosekingdom.rosekingdom.Core.Commands.ResourcePackTesting;
+import com.rosekingdom.rosekingdom.Core.Commands.TimersChecker;
 import com.rosekingdom.rosekingdom.Core.Utils.Message;
 import com.rosekingdom.rosekingdom.Economy.Commands.Coins;
 import com.rosekingdom.rosekingdom.Economy.Commands.Store;
@@ -50,6 +55,7 @@ public class CommandManager implements TabExecutor {
         addCommand(new Store());
         addCommand(new Ban());
         addCommand(new Vanish());
+        addCommand(new TimersChecker());
     }
 
     public List<CommandRK> getCommands() {
